requirements.txt
absl-py==2.1.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.3.1
colorama==0.4.6
comm==0.1.4
comtypes==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.20.0
Flask==3.1.2
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.11.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.5.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kagglehub==0.3.13
keras==3.12.0
kiwisolver==1.4.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.0.8
narwhals==2.12.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.6
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
opencv-contrib-python==4.8.1.78
opencv-python==4.12.0.88
opt-einsum==3.3.0
optree==0.12.1
overrides==7.4.0
pandas==2.3.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.11.0
plotly==6.5.0
polars==1.35.2
polars-runtime-32==1.35.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==6.33.1
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
pygame==2.6.0
Pygments==2.16.1
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2025.2
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==80.9.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.6.0
tinycss2==1.2.1
torch==2.9.1
torchvision==0.24.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.11.2
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.232
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.9
wordcloud==1.9.4
wrapt==1.16.0
